Import custom style overrides
The default Octopress theme imports a custom scss file at the end of its
screen.scss file that allows for people to override a theme's styles
with their own, custom styling. This commit takes that cue and
implements it for the Slash theme, importing the same file to allow for
overrides, making the theme more flexible.

More information on the default theme's custom style capabilities can be
found at:

  http://octopress.org/docs/theme/styles/