Use `h1` elements for post and page titles
The `h1` element serves as a signal that machines have been told to
interpret as "this is what this page is all about", similarly to the
`title` tag. Since each page should be unique and have a distinct
purpose, putting the site title in an `h1` on every single page is not
ideal. We'd rather be be giving these machines a clearer signal as to
what our pages are about.

This commit changes the element used for the site title from an `h1`
element to a plain `div`, and bumps the post and page titles from `h2`
to `h1`. This has the added bonus of improving the hierarchy of headings
in the posts themselves, which are likely to start at `h2`.

This should improve search engine optimization as well as integration
with other third-party services, such as Readability.

Conflicts:
	source/_includes/article.html